+++ keyboards/gray_studio/think65/hotswap/hotswap.c
@@ -13,7 +13,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "think65.h"
+#include "hotswap.h"
 
 // Optional override functions below.
 // You can leave any or all of these undefined.
